+++ samila/functions.py
@@ -3,6 +3,7 @@
 
 import requests
 from .params import Projection, DEFAULT_PROJECTION, VALID_COLORS, NFT_STORAGE_API, NFT_STORAGE_SUCCESS_MESSAGE, OVERVIEW
+import pickle
 
 
 def float_range(start, stop, step):
@@ -120,3 +121,20 @@ def samila_help():
     """
     print(OVERVIEW)
     print("Repo : https://github.com/sepandhaghighi/samila")
+
+
+def isSimilarData(data, path2data, precision=10**-5):
+    """
+    Compare the data is the same with given file.
+
+    :param data: given data
+    :type data: list
+    :param path2data: comparing data path on storage
+    :type path2data: str
+    :param precision: comparing precision
+    :type precision: float
+    :return: True if they are the same
+    """
+    with open(path2data, "rb") as fp:
+        temp_data = pickle.load(fp)
+    return all(map(lambda x, y: abs(x - y) < precision, data, temp_data))