+++ test/function_test.py
@@ -3,6 +3,7 @@
 >>> import random
 >>> import math
 >>> from samila import GenerativeImage, Projection
+>>> from samila.functions import isSimilarData
 >>> import pickle
 >>> def f1(x,y):
 ...    result = random.uniform(-1,1) * x**2  - math.sin(y**2) + abs(y-x)
@@ -24,13 +25,9 @@ True
 >>> g.generate(seed=10, start=-2*math.pi, step=0.1, stop=math.pi/2)
 >>> g.seed
 10
->>> with open("test/test1_1_d1.pkl", "rb") as fp:
-...   temp_data = pickle.load(fp)
->>> temp_data == g.data1
+>>> isSimilarData(g.data1, "test/test1_1_d1.pkl")
 True
->>> with open("test/test1_1_d2.pkl", "rb") as fp:
-...   temp_data = pickle.load(fp)
->>> temp_data == g.data2
+>>> isSimilarData(g.data2, "test/test1_1_d2.pkl")
 True
 >>> g.plot()
 >>> g.plot(color='red')